requirements.txt
google-generativeai
openai
speechrecognition
pygame
pyaudio
pydub
gtts
pyttsx3
python-dotenv
gradio
